+++ lib/sierra-models/base.js
@@ -12,6 +12,7 @@ class SierraBase {
         return map
       }, {})
     }
+    this.parallelsCache = this._cacheParallels()
   }
 
   isNyplRecord () {
@@ -113,6 +114,62 @@ class SierraBase {
     return parallelValues
   }
 
+  _cacheParallels () {
+    const parallels = this.varField(880, null, { excludedSubfields: [] })
+    return parallels.reduce((acc, parallelField) => {
+      const subfield6 = parallelField.subfieldMap[6]
+      // concat all subfields of parallel together???
+      const parallelContent = Object.keys(parallelField.subfieldMap).reduce((content, subfield) => {
+        if (subfield !== '6') {
+          return content + parallelField.subfieldMap[subfield]
+        } else return content
+      }, '')
+      // return [parallelFieldLink]: [value]
+      const { tag, number } = this._parseParallelFieldLink(subfield6)
+      return { ...acc, [tag + '-' + number]: parallelContent }
+    }, {})
+  }
+
+  _extractSubfield6 (field) {
+    const subfields = field.subfields || field.subFields
+    const subfield6 = subfields.find((sub) => sub.tag === '6')
+    const content = subfield6?.content
+    if (content) {
+      const parsed = this._parseParallelFieldLink(content)
+      return parsed
+    }
+  }
+
+  // primaries is array of fields returned from
+  //   SierraBase._extractValuesFromVarfields.
+  _addParallels (primaries, parallels) {
+    const primariesWithAndWithoutMatches = primaries.map((field, i) => {
+      // get subfield6
+      const subfieldSix = this._extractSubfield6(field)
+      const subfieldSixSuffix = subfieldSix?.number
+      // if its in table,
+      if (this.parallelsCache[subfieldSixSuffix]) {
+        //    add parallel to the primary
+        field.parallel = this.parallelsCache[subfieldSixSuffix]
+        //    remove parallel from table, or find another way to signify that its an orphan
+        this.parallelsCache[subfieldSixSuffix] = null
+      }
+      // return primary
+      return field
+    })
+    const orphanParallels = parallels.map((p) => {
+      // get subfieldSixSuffix
+
+      const subfieldSix = this._extractSubfield6(p)
+      const subfieldSixSuffix = subfieldSix?.number
+      // if that subfield 6 remains in lookup table
+      if (this.parallelsCache[subfieldSixSuffix]) {
+        return { parallel: p }
+      } else return null
+    }).filter((p) => p)
+    return primariesWithAndWithoutMatches.concat(orphanParallels)
+  }
+
   // Return content for given marc tag
   // if subfields given, returns content of those fields joined (or as a hash if opts.tagSubfields is truthy)
   // Options include:
@@ -121,8 +178,12 @@ class SierraBase {
   //  * preFilter: Optional function to filter out matching marc blocks before extracting value. Useful for checking atypical properties like `ind1`.
   //  * excludeSubfields: Optional array identifying subfields that should be excluded (e.g. ['0', '6'])
   varField (marc, subfields, opts) {
+    // _varFieldByMarcTag returns array of raw marc fields filtered by marc tag:
+    //  [{fieldTag: 'a', marcTag: '123', content: 'the thing'}]
     const fields = this._varFieldByMarcTag(marc)
-    // does this need to return an array if we're not using its return value anymore?
+    // extractValuesFromVarfields returns array with massaged values per subfield
+    //  [{value: 'subfields concatenated', subfieldMap: { a: 'subfields', b: 'concatenated'}}]
+    //  for 880 marcTag fields, you end up with the full parallel link and the subfield map
     return this._extractValuesFromVarFields(fields, subfields, opts)
   }
 
@@ -143,10 +204,12 @@ class SierraBase {
     return this._vars().filter((f) => !f.marcTag && f.fieldTag === `${fieldTag}`)
   }
 
+  // returns array of raw field data, filtered by marc tag
   _varFieldByMarcTag (marc) {
     return this._vars().filter((f) => f.marcTag === `${marc}`)
   }
 
+  // returns array of objects - {value: 'value', subfieldMap: {a:'val', 6:'parallel link'}}
   _extractValuesFromVarFields (varFields, subfields, opts) {
     opts = opts || {}
     opts = Object.assign({
@@ -157,6 +220,9 @@ class SierraBase {
     }, opts)
 
     if (Array.isArray(varFields) && varFields.length) {
+      // let allSubfields = varFields.every((field) => {
+      //   return field.marcTag === 880
+      // })
       const vals = varFields.filter(opts.preFilter).map((f) => {
         // sometimes there's a case error...
         const unfilteredFields = f.subFields || f.subfields
